refactor: break manuscriptReview <-> manuscriptFix static import cycle (#5918)
Move the review-comment store (schema, sanitizers, persistence, accessors)
into pipeline/manuscriptComments.js. manuscriptReview.js keeps seed
orchestration and re-exports the store API; manuscriptFix.js reads comments
through the leaf. Drops the #5918 baseline entry from the import-cycle
ratchet.
