+++ server/services/serviceImportCycles.test.js
@@ -79,8 +79,6 @@ const KNOWN_CYCLIC_COMPONENTS = [
   },
   // #5917 — caption resolution and subject derivation import each other.
   { issue: 5917, members: ['loraDatasetCaption.js', 'loraDatasetGenerate.js'] },
-  // #5918 — the fixer reaches back for the review-comment store accessors.
-  { issue: 5918, members: ['pipeline/manuscriptFix.js', 'pipeline/manuscriptReview.js'] },
   // #5919 — the receive path reaches subscription state back through the barrel.
   { issue: 5919, members: ['sharing/peerSync.js', 'sharing/peerSyncReceive.js'] },
 ];
